requirements.txt
django
djangorestframework
python-dotenv
mysqlclient
celery
redis
djangorestframework-simplejwt